Move paginator to file
